Refactor: Set the Android app module namespace
The namespace for the Android app module has been updated to `com.samuelokello.smartattendance.android`. This change is required for newer versions of the Android Gradle Plugin, which now mandate specifying a namespace in the module's `build.gradle.kts` file. Additionally the dependency `androidx.core.splashscreen` is added to enable splash screen capabilities.